requirements.txt
langchain
langchain-core
langchain-google-genai
streamlit
python-dotenv